UPDATES::
Added SyncFireStore for synchronous requests firestore.

Added ownerId and Id in Telemetry.The ownerId is the user email and Id is a unique Id for the Telemetry that is created by hashing the telemetry values.

Added Private/Public Keys Encryption in AES that are stored in firestore.

Email is Encrypted instead of hash, but passwords are still hashed.

Changed SSN to Id that represents UAE ID and added a validation for it in UserValidation called IdValidation function.

FIXES::
1. When Application is uninstalled but the user is still active in the firestore, still he can login into the application.

2. When Adding a telemetry the button will keep loading until the telemetry is added in the firestore.

3. When Setting the item to be auctioned, a toast will popup telling that the telemetry is now being auctioned.